% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{af_colour_palettes}
\alias{af_colour_palettes}
\title{Analysis Function colour palettes}
\format{
A character list
}
\source{
\href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{Government Analysis Function Colours Guidance}
}
\usage{
af_colour_palettes
}
\description{
A list grouping colours into palettes.
}
\keyword{datasets}
