% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{available_palettes}
\alias{available_palettes}
\alias{available_palette_types}
\title{Available palettes and palette types}
\usage{
available_palettes(palette_type, error_call = rlang::caller_env())

available_palette_types()
}
\arguments{
\item{palette_type}{String.}

\item{error_call}{Environment to reference in error messages.}
}
\value{
A character vector of available palettes or palette types.
}
\description{
Available palettes and palette types
}
\examples{
available_palette_types()
available_palettes("sg")

}
