% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{check_palette}
\alias{check_palette}
\alias{check_palette_type}
\title{Check palettes and palette types}
\usage{
check_palette(palette_type, palette)

check_palette_type(palette_type)
}
\arguments{
\item{palette_type, palette}{String.}
}
\value{
The value being checked is returned invisibly if the check is
successful. Otherwise the function will return an error.
}
\description{
Checks values against those available in
`available_palette_types()` and `available_palettes()`.
}
\examples{
check_palette_type("sg")
check_palette("sg", "main")

}
