% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_table.R
\name{colour_table}
\alias{colour_table}
\title{Create \code{gt} table of colour palette}
\usage{
colour_table(pal)
}
\arguments{
\item{pal}{Named vector of colour palette;
e.g. \code{sgplot::sg_colour_palettes[["main"]]}}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Create \code{gt} table of colour palette
}
