% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_sgplot.R
\name{disable_sgplot}
\alias{disable_sgplot}
\title{Revert sgplot defaults.}
\usage{
disable_sgplot()
}
\description{
Disable use of sgplot theme, colour palettes and geom aesthetic
defaults.
}
\examples{
library(ggplot2)

d <- subset(mpg, manufacturer == "ford")

use_sgplot()
ggplot(d, aes(x = model, fill = class)) + geom_bar()

disable_sgplot()

ggplot(d, aes(x = model, fill = class)) + geom_bar()

}
