% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_to_inch}
\alias{mm_to_inch}
\title{Convert millimetres to inches}
\usage{
mm_to_inch(x)
}
\arguments{
\item{x}{Numeric value in millimetres}
}
\value{
A numerical value in inches
}
\description{
Convert millimetres to inches
}
\examples{
mm_to_inch(100)

}
