% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_discrete_scs.R
\name{scale_colour_discrete_scs}
\alias{scale_colour_discrete_scs}
\title{Discrete colour scales for SCS}
\usage{
scale_colour_discrete_scs(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use from `scs_colour_palettes`; e.g. main,
sequential, focus.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed}

\item{...}{Additional arguments passed to scale type}
}
\description{
Discrete colour scales for SCS
}
\examples{
library(ggplot2)
qplot(mpg, wt, data = mtcars, colour = as.factor(cyl)) +
 scale_colour_discrete_scs()

}
