% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_discrete_sg.R
\name{scale_colour_discrete_sg}
\alias{scale_colour_discrete_sg}
\title{Discrete colour scales for Scottish Government plots}
\usage{
scale_colour_discrete_sg(
  palette = "main",
  palette_type = c("sg", "af"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Name of palette to use; e.g. "main", "sequential", "focus."
Default value is "main".}

\item{palette_type}{Either "sg" to use Scottish Government palettes, or "af"
to use Analysis Function palettes. Defaults to "sg".}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed.}

\item{...}{Additional arguments passed to scale type.}
}
\description{
Discrete colour scales for Scottish Government plots
}
\examples{
library(ggplot2)
library(dplyr)

economics_long \%>\%
  filter(variable \%in\% c("psavert", "uempmed")) \%>\%
  ggplot(aes(x = date, y = value, colour = variable)) +
  geom_line(linewidth = 1) +
  scale_colour_discrete_sg()

}
