% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_continuous_sg.R
\name{scale_fill_continuous_sg}
\alias{scale_fill_continuous_sg}
\title{Continuous colour fill scales for Scottish Government plots}
\usage{
scale_fill_continuous_sg(
  palette = "sequential",
  reverse = FALSE,
  na_colour = "grey50",
  guide = "colourbar",
  ...
)
}
\arguments{
\item{palette}{Name of palette to use from `sg_colour_palettes`; e.g. main,
sequential, focus. Default value is 'sequential'.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed.}

\item{na_colour}{Colour to set for missing values.}

\item{guide}{A name or function used to create guide. Default is "colourbar".}

\item{...}{Additional arguments passed to scale type.}
}
\description{
Continuous colour fill scales for Scottish Government plots
}
\examples{
library(ggplot2)

ggplot(faithfuld, aes(x = waiting, y = eruptions, fill = density)) +
  geom_raster() +
  scale_fill_continuous_sg()

}
