% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_discrete_scs.R
\name{scale_fill_discrete_scs}
\alias{scale_fill_discrete_scs}
\title{Discrete colour fill scales for SCS}
\usage{
scale_fill_discrete_scs(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use from `scs_colour_palettes`; e.g. main,
sequential, focus.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed}

\item{...}{Additional arguments passed to scale type}
}
\description{
Discrete colour fill scales for SCS
}
\examples{
library(ggplot2)
library(dplyr)
mpg \%>\%
 filter(substr(manufacturer, 1, 1) \%in\% c("a", "c", "f")) \%>\%
 ggplot(aes(x = manufacturer, y = cty, fill = manufacturer)) +
 geom_col() +
 scale_fill_discrete_scs()

}
