% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_discrete_sg.R
\name{scale_fill_discrete_sg}
\alias{scale_fill_discrete_sg}
\title{Discrete colour fill scales for Scottish Government plots}
\usage{
scale_fill_discrete_sg(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use from `sg_colour_palettes`; e.g. main,
sequential, focus. Default value is 'main'.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed.}

\item{...}{Additional arguments passed to scale type.}
}
\description{
Discrete colour fill scales for Scottish Government plots
}
\examples{
library(ggplot2)

d <- subset(mpg, manufacturer == "ford")

ggplot(d, aes(x = class, fill = class)) +
 geom_bar() +
 scale_fill_discrete_sg()

}
