% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scs_colour_palettes}
\alias{scs_colour_palettes}
\title{SCS colour names and hex codes.}
\format{
A character list
}
\usage{
scs_colour_palettes
}
\description{
A list grouping SCS colours into palettes.
}
\keyword{datasets}
