% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scs_colour_values}
\alias{scs_colour_values}
\title{SCS colour names and hex codes.}
\format{
A character vector
}
\source{
https://gss.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/
}
\usage{
scs_colour_values
}
\description{
A vector containing SCS colour names and their corresponding
hex code.
}
\keyword{datasets}
