% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs_palette.R
\name{scs_palette}
\alias{scs_palette}
\title{Return function to use SCS colour palette}
\usage{
scs_palette(palette = "main", reverse = FALSE, colour_names = FALSE)
}
\arguments{
\item{palette}{Name of palette from `scs_colour_palettes`}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed}

\item{colour_names}{Boolean value to indicate whether colour names should be
included}
}
\description{
Return function to use SCS colour palette
}
