% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs_theme.R
\name{scs_theme}
\alias{scs_theme}
\title{Scottish Contact Survey theme for ggplot2 charts.}
\usage{
scs_theme()
}
\description{
ggplot2 theme for reporting of Scottish Government Covid-19
Scottish Contact Survey (SCS) analysis.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = manufacturer, y = cty)) +
   geom_col() +
   scs_theme()

}
