% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sg_colour_values}
\alias{sg_colour_values}
\title{Scottish Government colour names and hex codes}
\format{
A character vector
}
\source{
\href{https://designsystem.gov.scot/guidance/charts/data-visualisation-colour-palettes}{Scottish Government Design System}
}
\usage{
sg_colour_values
}
\description{
A vector containing colour names and their corresponding
hex code.
}
\keyword{datasets}
