% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_theme.R
\name{sg_theme}
\alias{sg_theme}
\title{Scottish Government theme for ggplot2 charts.}
\usage{
sg_theme()
}
\description{
ggplot2 theme for Scottish Government plots.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = manufacturer, y = cty)) +
   geom_col() +
   sg_theme()

}
