% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sss_colour_palettes}
\alias{sss_colour_palettes}
\title{Social Security Scotland colour palettes}
\format{
A character list
}
\usage{
sss_colour_palettes
}
\description{
A list grouping colours into palettes.
}
\keyword{datasets}
