% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sss_colour_values}
\alias{sss_colour_values}
\title{Social Security Scotland colour names and hex codes}
\format{
A character vector
}
\usage{
sss_colour_values
}
\description{
A vector containing colour names and their corresponding
hex code.
}
\keyword{datasets}
