% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_discrete_sg.R
\name{scale_colour_discrete_sg}
\alias{scale_colour_discrete_sg}
\title{Discrete colour scales for Scottish Government plots}
\usage{
scale_colour_discrete_sg(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use from `sg_colour_palettes`; e.g. main,
sequential, focus.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed}

\item{...}{Additional arguments passed to scale type}
}
\description{
Discrete colour scales for Scottish Government plots
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = wt, colour = as.factor(cyl))) +
  geom_point() +
  scale_colour_discrete_sg() +
  theme_sg()

}
