% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_discrete_sg.R
\name{scale_fill_discrete_sg}
\alias{scale_fill_discrete_sg}
\title{Discrete colour fill scales for Scottish Government plots}
\usage{
scale_fill_discrete_sg(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use from `sg_colour_palettes`; e.g. main,
sequential, focus.}

\item{reverse}{Boolean value to indicate whether the palette should be
reversed}

\item{...}{Additional arguments passed to scale type}
}
\description{
Discrete colour fill scales for Scottish Government plots
}
\examples{
library(ggplot2)
library(dplyr)
mpg \%>\%
  filter(substr(manufacturer, 1, 1) \%in\% c("a", "c", "f")) \%>\%
  ggplot(aes(x = manufacturer, y = cty, fill = manufacturer)) +
  geom_col() +
  scale_fill_discrete_sg() +
  theme_sg()

}
