% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sg_colour_palettes}
\alias{sg_colour_palettes}
\title{Scottish Government colour palettes}
\format{
A character list
}
\source{
\href{https://designsystem.gov.scot/guidance/charts/data-visualisation-colour-palettes}{Scottish Government Design System}
}
\usage{
sg_colour_palettes
}
\description{
A list grouping colours into palettes.
}
\keyword{datasets}
