% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sg_colour_values}
\alias{sg_colour_values}
\title{Scottish Government colour names and hex codes}
\format{
A character vector
}
\source{
\href{https://analysisfunction.civilservice.gov.uk/policy-store/
data-visualisation-colours-in-charts/}{Government Analysis Function Colours
Guidance}
}
\usage{
sg_colour_values
}
\description{
A vector containing colour names and their corresponding
hex code.
}
\keyword{datasets}
