% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgplot-package.R
\docType{package}
\name{sgplot-package}
\alias{sgplot}
\alias{sgplot-package}
\title{sgplot: Graphic Styles and Colours for Scottish Government Plots}
\description{
A 'ggplot2' theme and colour palettes to create accessible data visualisations in the Scottish Government.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ScotGovAnalysis.github.io/sgplot/}
  \item \url{https://github.com/ScotGovAnalysis/sgplot}
  \item Report bugs at \url{https://github.com/ScotGovAnalysis/sgplot/issues}
}

}
\author{
\strong{Maintainer}: Alice Hannah \email{alice.hannah@gov.scot}

Other contributors:
\itemize{
  \item Scottish Government \email{statistics.enquiries@gov.scot} [copyright holder, funder]
}

}
\keyword{internal}
