% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sss_colour_values}
\alias{sss_colour_values}
\alias{sss_colour_palettes}
\title{Social Security Scotland colours and palettes}
\format{
An object of class \code{character} of length 12.

An object of class \code{list} of length 4.
}
\source{
Contact the \href{mailto:MI@socialsecurity.gov.scot}{Social Security Scotland Statistics mailbox}
with any queries about these colours and palettes.
}
\usage{
sss_colour_values

sss_colour_palettes
}
\description{
\itemize{
\item \code{sss_colour_values} is a vector containing colour names and their corresponding
hex codes.
\item \code{sss_colour_palettes} is a list grouping colours into palettes.
}
}
\keyword{datasets}
