% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sgplot.R
\name{use_sgplot}
\alias{use_sgplot}
\title{Use \code{sgplot} defaults.}
\usage{
use_sgplot(default_colour = sgplot::sg_colour_values["dark-blue"], ...)
}
\arguments{
\item{default_colour}{Default colour/fill for geoms. Default value is
'dark-blue' from \code{sgplot::sg_colour_values}.}

\item{...}{Arguments passed to \code{theme_sg}.}
}
\description{
Set \code{sgplot} theme, colour palette and geom aesthetic
defaults for ggplot2 charts.
}
\examples{
library(ggplot2)

d <- subset(mpg, manufacturer == "ford")

ggplot(d, aes(x = model)) + geom_bar()
ggplot(d, aes(x = model, fill = class)) + geom_bar()

use_sgplot()

ggplot(d, aes(x = model)) + geom_bar()
ggplot(d, aes(x = model, fill = class, colour = class)) + geom_bar()

}
