% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sgplot.R
\name{use_sgplot}
\alias{use_sgplot}
\title{Use \code{sgplot} as default.}
\usage{
use_sgplot(default_col = sg_colour_values["dark-blue"])
}
\arguments{
\item{default_col}{Default colour for geoms. Default value is 'dark-blue'
from \code{sgplot::sg_colour_values}.}
}
\description{
Set \code{sgplot} theme and geom aesthetics as default for
ggplot2 charts.
}
\examples{
use_sgplot()

}
